% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{get_domain_extension}
\alias{get_domain_extension}
\title{Get domain extension (gTLD or ccTLD) from URL or domain name}
\usage{
get_domain_extension(url)
}
\arguments{
\item{url}{The URL or domain.}
}
\value{
Return normal domain name in text
}
\description{
Get domain extension from a URL or domain.
}
\examples{
\dontrun{
get_domain_extension("example.com")
}

}
