% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{get_normal_text}
\alias{get_normal_text}
\title{Get Normat Text from a punycode}
\usage{
get_normal_text(domain)
}
\arguments{
\item{domain}{The punycode domain name}
}
\value{
Return normal domain name in text
}
\description{
Get Normat Text from a punycode for domain name.
}
\examples{
\dontrun{
get_normal_text("xn--tst-qla.de")
}

}
