% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{get_punycode}
\alias{get_punycode}
\title{Get Punycode for domain name}
\usage{
get_punycode(domain)
}
\arguments{
\item{domain}{domain name to get punycode for}
}
\value{
Return the converted punycode of the domain
}
\description{
Get Punycode for domain name.
}
\examples{
\dontrun{
get_punycode("täst.de")
}

}
