% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookup}
\alias{lookup}
\title{Lookup for WHOIS information}
\usage{
lookup(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the WHOIS information about the domain
}
\description{
Lookup for a comprehensive of the WHOIS information
}
\examples{
\dontrun{
lookup("example.com")
}

}
