% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupAdmin}
\alias{lookupAdmin}
\title{Lookup for admin information}
\usage{
lookupAdmin(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the admin information of the domain
}
\description{
Lookup for admin information
}
\examples{
\dontrun{
lookupAdmin("example.com")
}

}
