% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupBilling}
\alias{lookupBilling}
\title{Lookup for billing information}
\usage{
lookupBilling(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the billing information of the domain
}
\description{
Lookup for billing information
}
\examples{
\dontrun{
lookupBilling("example.com")
}

}
