% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupNameservers}
\alias{lookupNameservers}
\title{Lookup for nameservers information}
\usage{
lookupNameservers(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the nameservers information of the domain
}
\description{
Lookup for nameservers information
}
\examples{
\dontrun{
lookupNameservers("example.com")
}

}
