% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupRegistrant}
\alias{lookupRegistrant}
\title{Lookup for registrant information}
\usage{
lookupRegistrant(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the registrant information of the domain
}
\description{
Lookup for registrant information
}
\examples{
\dontrun{
lookupRegistrant("example.com")
}

}
