% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupRegistrar}
\alias{lookupRegistrar}
\title{Lookup for registrar information}
\usage{
lookupRegistrar(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the registrar information of the domain
}
\description{
Lookup for registrar information
}
\examples{
\dontrun{
lookupRegistrar("example.com")
}

}
