% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupTech}
\alias{lookupTech}
\title{Lookup for tech information}
\usage{
lookupTech(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the tech information of the domain
}
\description{
Lookup for tech information
}
\examples{
\dontrun{
lookupTech("example.com")
}

}
