% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{lookupWhoisServer}
\alias{lookupWhoisServer}
\title{Lookup for whois server information}
\usage{
lookupWhoisServer(domain)
}
\arguments{
\item{domain}{domain name to lookup for}
}
\value{
Return the whois server information of the domain
}
\description{
Lookup for whois server information
}
\examples{
\dontrun{
lookupWhoisServer("example.com")
}

}
