% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP2whois.R
\name{set_api_key}
\alias{set_api_key}
\title{Set IP2Location.io API key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{IP2Location.io API key}
}
\description{
Set IP2Location.io API key for lookup. Free API key can be obtained from <https://www.ip2location.io/sign-up?ref=1/>
}
\examples{
\dontrun{
set_api_key("YOUR_API_KEY")
}

}
